+++ axum-extra/src/extract/mod.rs
@@ -40,4 +40,8 @@ pub use self::multipart::Multipart;
 #[doc(no_inline)]
 pub use crate::json_lines::JsonLines;
 
+#[cfg(feature = "typed-header")]
+#[doc(no_inline)]
+pub use crate::typed_header::TypedHeader;
+
 pub use self::with_rejection::WithRejection;