+++ axum-extra/src/lib.rs
@@ -21,6 +21,7 @@
 //! `protobuf` | Enables the `Protobuf` extractor and response | No
 //! `query` | Enables the `Query` extractor | No
 //! `typed-routing` | Enables the `TypedPath` routing utilities | No
+//! `typed-header` | Enables the `TypedHeader` extractor and response  | No
 //!
 //! [`axum`]: https://crates.io/crates/axum
 
@@ -78,6 +79,9 @@ pub mod routing;
 
 #[cfg(feature = "json-lines")]
 pub mod json_lines;
+
+#[cfg(feature = "typed-header")]
+pub mod typed_header;
 
 #[cfg(feature = "protobuf")]
 pub mod protobuf;