+++ axum-macros/src/lib.rs
@@ -71,9 +71,12 @@ use from_request::Trait::{FromRequest, FromRequestParts};
 /// ```
 /// use axum_macros::FromRequest;
 /// use axum::{
-///     extract::{Extension, TypedHeader},
-///     headers::ContentType,
+///     extract::Extension,
 ///     body::Bytes,
+/// };
+/// use axum_extra::typed_header::{
+///     TypedHeader,
+///     headers::ContentType,
 /// };
 ///
 /// #[derive(FromRequest)]
@@ -116,9 +119,12 @@ use from_request::Trait::{FromRequest, FromRequestParts};
 /// ```
 /// use axum_macros::FromRequest;
 /// use axum::{
-///     extract::{Extension, TypedHeader},
-///     headers::ContentType,
+///     extract::Extension,
 ///     body::Bytes,
+/// };
+/// use axum_extra::typed_header::{
+///     TypedHeader,
+///     headers::ContentType,
 /// };
 ///
 /// #[derive(FromRequest)]
@@ -158,8 +164,9 @@ use from_request::Trait::{FromRequest, FromRequestParts};
 ///
 /// ```
 /// use axum_macros::FromRequest;
-/// use axum::{
-///     extract::{TypedHeader, rejection::TypedHeaderRejection},
+/// use axum_extra::typed_header::{
+///     TypedHeader,
+///     TypedHeaderRejection,
 ///     headers::{ContentType, UserAgent},
 /// };
 ///
@@ -368,7 +375,10 @@ pub fn derive_from_request(item: TokenStream) -> TokenStream {
 /// ```
 /// use axum_macros::FromRequestParts;
 /// use axum::{
-///     extract::{Query, TypedHeader},
+///     extract::Query,
+/// };
+/// use axum_extra::typed_header::{
+///     TypedHeader,
 ///     headers::ContentType,
 /// };
 /// use std::collections::HashMap;